Rename proyection module to proyeccion
